% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gg_info.R
\name{find_gg_info}
\alias{find_gg_info}
\title{Find build information of a ggplot}
\usage{
find_gg_info(graph)
}
\arguments{
\item{graph}{A ggplot to analyze.}
}
\value{
A list with two entries:
\itemize{
\item \code{main_data}: the data used in \code{ggplot()}.
\item \code{layers}: for each layer the aesthetic mapping and the data used (which
can be \code{'inherited'}).
}
}
\description{
To customize a plot made in a package, adding new layers, it is useful to
see the data and mappings used. This function returns a list with both.
}
\examples{
g <- ggplot(mtcars, aes(mpg, hp)) + geom_point(aes(color = gear))
find_gg_info(g)

}
