% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_acf.R, R/ggvar_ccf.R
\name{ggvar_acf}
\alias{ggvar_acf}
\alias{ggvar_ccf}
\title{Plot autocorrelation (and similars) of dataset}
\usage{
ggvar_acf(
  x,
  series = NULL,
  lag.max = NULL,
  type = "correlation",
  graph_type = "segment",
  args_geom = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(),
  args_facet = list(),
  ci = 0.95,
  ...
)

ggvar_ccf(
  x,
  series = NULL,
  lag.max = NULL,
  type = "correlation",
  graph_type = "segment",
  args_geom = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(),
  args_facet = list(),
  facet_type = "ggplot",
  ci = 0.95,
  ...
)
}
\arguments{
\item{x}{A dataset (object coercible to data.frame) or a "varest" object to
get residuals from.}

\item{series}{A character vector with series (column names) to consider. Defaults to all (\code{NULL}).}

\item{lag.max}{The number of lags used to calculate the ACF, passed to
\link[stats:acf]{acf}.}

\item{type}{The type of ACF to be computed. Can be either "correlation",
"covariance", or "partial". Passed to \link[stats:acf]{acf}.}

\item{graph_type}{The ggplot geom used to create the plot: \code{'segment'}, for \link[ggplot2:geom_segment]{ggplot2::geom_segment} or \code{'area'}, for \link[ggplot2:geom_ribbon]{ggplot2::geom_area}.}

\item{args_geom}{Arguments passed to the chosen 'geom'.}

\item{args_ribbon}{Additional arguments passed to \link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon}.}

\item{args_hline}{Additional arguments passed to \link[ggplot2:geom_abline]{ggplot2::geom_hline}.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{ci}{The level of confidence for the ACF confidence interval. Set to
\code{FALSE} to omit the \link[ggplot2:geom_ribbon]{geom_ribbon}.}

\item{...}{Arguments passed to \verb{varr:::setup_acf_setup or varr:::setup_ccf_setup}, the generic function that formats \code{x} into a 'graphable' format. Use them only if you have created a method for some unsupported class of \code{x}.Pass additional arguments to \link[stats:acf]{stats::acf} here.}

\item{facet_type}{The facet 'engine' to be used. \code{'ggplot2'} for \link[ggplot2:facet_grid]{ggplot2::facet_grid}, \code{'ggh4x'} for \link[ggh4x:facet_grid2]{ggh4x::facet_grid2}.}
}
\value{
An object of class \code{'ggplot'}.
}
\description{
\code{ggvar_acf} plots the auto-correlations (and similars) call for every
series, using \link[ggplot2:facet_wrap]{facet_wrap}. \code{ggvar_ccf} plots all the
cross-correlations (and similars) between the series, in a grid.
}
\examples{
ggvar_acf(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_ccf(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_acf(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
