% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_irf}
\alias{ggvar_irf}
\title{Plot VAR impulse response functions}
\usage{
ggvar_irf(
  x,
  series_imp = NULL,
  series_resp = NULL,
  n.ahead = 10,
  ci = 0.95,
  args_line = list(),
  args_hline = list(),
  args_ribbon = list(fill = NA, linetype = 2, color = "blue"),
  args_facet = list(),
  facet_type = "ggplot",
  ...
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars:irf]{irf}, or, directly, a
"varirf" object.}

\item{series_imp}{A character vector with variables to consider for the
impulses. Defaults to all (\code{NULL}).}

\item{series_resp}{A character vector with variables to consider for the
responses. Defaults to all (\code{NULL}).}

\item{n.ahead}{An integer. The size of the forecast horizon, passed to
\link[vars:irf]{irf}. Unused if \code{x} is of class varirf.}

\item{ci}{The confidence level for the confidence interval. Set to \code{FALSE} to omit. Passed to \link[vars:irf]{irf}.}

\item{args_line}{Additional arguments passed to \link[ggplot2:geom_path]{geom_line}.}

\item{args_hline}{Additional arguments passed to \link[ggplot2:geom_abline]{geom_hline}.}

\item{args_ribbon}{Additional arguments passed to \link[ggplot2:geom_ribbon]{geom_ribbon}.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{facet_type}{The facet 'engine' to be used. \code{'ggplot2'} for \link[ggplot2:facet_grid]{ggplot2::facet_grid}, \code{'ggh4x'} for \link[ggh4x:facet_grid2]{ggh4x::facet_grid2}.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Plots the result of a \link[vars:irf]{irf} call.
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{find_gg_info}.

See \link{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function varr:::irf_setup. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to \link[vars:irf]{vars::irf}.
\item Class \code{'varirf'}: passed to nothing.
}
}
}
\examples{
ggvar_irf(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
\seealso{
Other VAR output plots: 
\code{\link{ggvar_fevd}()}
}
\concept{VAR output plots}
