% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varr-package.R
\docType{package}
\name{mtsdesc-package}
\alias{mtsdesc}
\alias{mtsdesc-package}
\title{mtsdesc: Plots and Tables for Describing Multivariate Time Series}
\description{
Describe multivariate time series with tables and ggplots. mtsdesc provides commonly needed utilities when working with a set of time series, i.e. multi-panel plots and multi-column summary- and testing-tables. The package also has methods to work with residuals of fitter values of multivariate time series models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/ricardo-semiao/mtsdesc}
  \item \url{https://ricardo-semiao.github.io/mtsdesc/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/mtsdesc/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
