% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varr-package.R
\docType{package}
\name{varr-package}
\alias{varr}
\alias{varr-package}
\title{varr: Utility Functions for VAR Analisys}
\description{
Functions for commonly needed ggplot graphs for Vector Auto-Regressive models, like IRF's, FEVD's, predict graphs, and VAR residual analysis. Soon to contain functions for common tests and statistics, and other VAR-like objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/ricardo-semiao/varr}
  \item \url{https://ricardo-semiao.github.io/varr/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/varr/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
