% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_acf.R, R/ggvar_ccf.R
\name{ggvar_acf}
\alias{ggvar_acf}
\alias{ggvar_ccf}
\title{Plot autocorrelation and similar}
\usage{
ggvar_acf(
  x,
  series = NULL,
  lag.max = NULL,
  type = "correlation",
  graph_type = "segment",
  args_type = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(yintercept = 0),
  args_labs = list(),
  args_facet = list(),
  ci = 0.95,
  ...
)

ggvar_ccf(
  x,
  series = NULL,
  lag.max = NULL,
  type = "correlation",
  graph_type = "segment",
  args_type = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(yintercept = 0),
  args_labs = list(),
  args_facet = list(),
  facet_type = "ggplot",
  ci = 0.95,
  ...
)
}
\arguments{
\item{x}{A dataset (object coercible to "data.frame") or a "varest" object to
get residuals from.}

\item{series}{A character vector with series (variables) to  consider. Defaults to all (\code{NULL}).}

\item{lag.max}{The number of lags used to calculate the ACF, passed to
\link[stats:acf]{acf}.}

\item{type}{The type of ACF to be computed. Can be either "correlation",
"covariance", or "partial". Passed to \link[stats:acf]{acf}.}

\item{graph_type}{The ggplot geom used to create the plot: {texts_vec}.}

\item{args_type}{Arguments passed to the 'geom' chosen in \code{graph_type}.}

\item{args_ribbon, args_hline}{Additional arguments passed to \link[ggplot2:geom_ribbon]{geom_ribbon} and \link[ggplot2:geom_abline]{geom_hline} (respectively). See more in the 'Customization' section.}

\item{args_labs}{Additional arguments passed to \link[ggplot2:labs]{labs}. If an empty list, will be changed to default values.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{ci}{The confidence level for the confidence interval. Set to \code{FALSE} to omit. Used in \link[ggplot2:geom_ribbon]{geom_ribbon}.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}

\item{facet_type}{The facet 'engine' to be used. \code{'ggplot2'} for \link[ggplot2:facet_grid]{ggplot2::facet_grid}, \code{'ggh4x'} for \link[ggh4x:facet_grid2]{ggh4x::facet_grid2}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Given a dataset or VAR model to get residuals from, \code{ggvar_acf} plots the
auto-correlations (and similar) call for every series. \code{ggvar_ccf} plots all
the cross-correlations (and similar) between the series, in a grid.
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

The \code{args_aes} is a list with '* = arguments to \verb{scale_*_manual} elements, where '*' represents the name of an aesthetic to apply to the data. View \link{vignette('ggplot2-specs', 'ggplot2')} to see the available aesthetics.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{find_gg_info}.

See \link{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function {texts_vec}. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to \link[stats:acf]{stats::acf}.
\item Class \code{'default'}: passed to \link[stats:acf]{stats::acf}.
}
}
}
\examples{
ggvar_acf(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_ccf(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_acf(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
\seealso{
Other general time series plots: 
\code{\link{ggvar_distribution}()},
\code{\link{ggvar_history}()}

Other model diagnostics plots: 
\code{\link{ggvar_dispersion}()},
\code{\link{ggvar_distribution}()},
\code{\link{ggvar_history}()},
\code{\link{ggvar_select}()},
\code{\link{ggvar_stability}()}
}
\concept{general time series plots}
\concept{model diagnostics plots}
