% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_distribution.R
\name{ggvar_distribution}
\alias{ggvar_distribution}
\title{Plot VAR residuals distribution}
\usage{
ggvar_distribution(
  x,
  series = NULL,
  plot_normal = TRUE,
  args_histogram = list(bins = 30),
  args_line = list(),
  args_labs = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{Either a "varest" object for plotting the residuals, or an dataset
(object coercible to data.frame) with numeric variables.}

\item{series}{A character vector with series (variables) to  consider. Defaults to all (\code{NULL}).}

\item{plot_normal}{Logical, whether or not a normal curve should be plotted.}

\item{args_histogram, args_line}{Additional arguments passed to \link[ggplot2:geom_histogram]{geom_histogram} and \link[ggplot2:geom_path]{geom_line} (respectively). See more in the 'Customization' section.}

\item{args_labs}{Additional arguments passed to \link[ggplot2:labs]{labs}. If an empty list, will be changed to default values.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Plots the histogram of the residuals of a VAR model, or of the variables in a
dataset, possibly overlapped with a normal curve.
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{get_gg_info}.

See \code{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function mtsdesc:::distribution_setup. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to nothing.
\item Class \code{'default'}: passed to nothing.
}
}
}
\examples{
ggvar_distribution(vars::VAR(freeny[-2]))

}
\seealso{
Other model diagnostics plots: 
\code{\link{ggvar_acf}()},
\code{\link{ggvar_dispersion}()},
\code{\link{ggvar_history}()},
\code{\link{ggvar_select}()},
\code{\link{ggvar_stability}()}

Other general time series plots: 
\code{\link{ggvar_acf}()},
\code{\link{ggvar_history}()}
}
\concept{general time series plots}
\concept{model diagnostics plots}
