% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_fit.R
\name{ggvar_fit}
\alias{ggvar_fit}
\title{Plot VAR Fitted Values}
\usage{
ggvar_fit(
  x,
  series = NULL,
  index = NULL,
  args_line = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{A "varest" object to get fitted values from.}

\item{series}{A character vector with series (column names) to consider. Defaults to all (\code{NULL}).}

\item{index}{A vector of labels to the x-axis, normally dates. Must have length equal to \code{x$obs}. Defaults to a integer sequence.}

\item{args_line}{Additional arguments passed to \link[ggplot2:geom_path]{ggplot2::geom_line}.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to \code{varr:::fit_setup}, the generic function that formats \code{x} into a 'graphable' format. Use them only if you have created a method for some unsupported class of \code{x}.}
}
\value{
An object of class \code{'ggplot'}.
}
\description{
Plots fitted values of a VAR model, versus the actual values.
\code{ggvar_fit} Plots each serie in a facet. \code{ggvar_fit_colored}
plots all in the same graph, each with a different color.
}
\examples{
x <- vars::VAR(freeny[-2])
ggvar_fit(x, args_facet = list(scales = "free_y"))

}
