% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_history.R
\name{ggvar_history}
\alias{ggvar_history}
\title{Plot historic values of dataset or VAR residuals}
\usage{
ggvar_history(
  x,
  series = NULL,
  index = NULL,
  faceted = TRUE,
  args_aes = list(),
  args_line = list(),
  args_labs = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{Either a "varest" object for plotting the residuals, or an dataset
(object coercible to data.frame) with numeric variables.}

\item{series}{A character vector with series (variables) to  consider. Defaults to all (\code{NULL}).}

\item{index}{A vector of labels to the x-axis, normally dates. Must have length equal to {texts_vec}. Defaults to a integer sequence.}

\item{faceted}{Should the graph be divided in facets?. If true, \code{args_aes} is ignored.}

\item{args_aes}{Defines aesthetics to differentiate the data. A named list of aesthetics names (\code{*}) -- arguments passed to \verb{ggplot2::scale_*_manual}. See more in the 'Customization' section.}

\item{args_line, facet_wrap}{Additional arguments passed to \link[ggplot2:geom_path]{geom_line} and \link[ggplot2:facet_wrap]{facet_wrap} (respectively). See more in the 'Customization' section.}

\item{args_labs}{Additional arguments passed to \link[ggplot2:labs]{labs}. If an empty list, will be changed to default values.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Plots the historic values of variables in a dataset, or residuals of a VAR
model. Can separate variables with facets or with a colors (see
\code{graph_type}).
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

The \code{args_aes} is a list with '* = arguments to \verb{scale_*_manual} \
elements, where '*' represents the name of an aesthetic to apply to the \
data. View \link{vignette('ggplot2-specs', 'ggplot2')} to see the available \
aesthetics.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{find_gg_info}.

See \link{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function {texts_vec}. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to nothing.
\item Class \code{'default'}: passed to nothing.
}
}
}
\examples{
ggvar_history(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_history(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
\seealso{
Other general time series plots: 
\code{\link{ggvar_acf}()},
\code{\link{ggvar_distribution}()}

Other historic values plots: 
\code{\link{ggvar_fit}()},
\code{\link{ggvar_predict}()}

Other model diagnostics plots: 
\code{\link{ggvar_acf}()},
\code{\link{ggvar_dispersion}()},
\code{\link{ggvar_distribution}()},
\code{\link{ggvar_select}()},
\code{\link{ggvar_stability}()}
}
\concept{general time series plots}
\concept{historic values plots}
\concept{model diagnostics plots}
