% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_history.R
\name{ggvar_history}
\alias{ggvar_history}
\title{Plot Values of Dataset or VAR Residuals}
\usage{
ggvar_history(
  x,
  series = NULL,
  index = NULL,
  graph_type = "faceted",
  args_line = list(),
  args_facet = list(),
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{Either a "varest" object for plotting the residuals, or an dataset
(object coercible to data.frame) with numeric variables.}

\item{series}{A character vector with series (column names) to consider. Defaults to all (\code{NULL}).}

\item{index}{A vector of labels to the x-axis, normally dates. Must have length equal to \code{x$obs} or \code{nrow(x)}. Defaults to a integer sequence.}

\item{graph_type}{The type of the plot, \code{'faceted'} or \code{'colored'}.}

\item{args_line}{Additional arguments passed to \link[ggplot2:geom_path]{ggplot2::geom_line}.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{colors}{A vector of colors for the 'geoms' and/or variables.
See \link{vignette('colors')}.}

\item{...}{Arguments passed to \code{varr:::history_setup}, the generic function that formats \code{x} into a 'graphable' format. Use them only if you have created a method for some unsupported class of \code{x}.}
}
\value{
An object of class \code{'ggplot'}.
}
\description{
Plots the historic values of variables in a dataset, or residuals of a VAR
model. \code{ggvar_values} Plots each series in a facet.
\code{ggvar_values_colored} plots all in the same graph, each with a
different color.
}
\examples{
ggvar_history(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_history(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
