% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_predict.R
\name{ggvar_predict}
\alias{ggvar_predict}
\title{Plot the Predicted Values of a VAR}
\usage{
ggvar_predict(
  x,
  series = NULL,
  index_ahead,
  index_behind = NULL,
  ci = 0.95,
  args_line = list(),
  args_ribbon = list(fill = NA, linetype = 2, color = "blue"),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{A "varest" object to get predictions from.}

\item{series}{A character vector with series (column names) to consider. Defaults to all (\code{NULL}).}

\item{index_ahead}{A vector of labels to the x-axis, normally dates. Applied
to the predicted portion of the graph. Its length will define the
prediction horizon.}

\item{index_behind}{A vector of labels to the x-axis, normally dates. Applied
to the original portion of the graph. Its length will define the 'past'
horizon. Leave as \code{NULL} to only plot predicted values.}

\item{ci}{The level of confidence for the prediction confidence interval. Set
to \code{FALSE} to omit. Passed to \link[stats:predict]{predict}.}

\item{args_line}{Additional arguments passed to \link[ggplot2:geom_path]{ggplot2::geom_line}.}

\item{args_ribbon}{Additional arguments passed to \link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon}.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to \code{varr:::predict_setup}, the generic function that formats \code{x} into a 'graphable' format. Use them only if you have created a method for some unsupported class of \code{x}.Pass additional arguments to \link[stats:predict]{stats::predict} here.}
}
\value{
An object of class \code{'ggplot'}.
}
\description{
Plots the result of a \link[vars:predict]{predict.varest} call. Has an option to
overlay it with the true variables, if provided a test dataset.
}
\examples{
x <- vars::VAR(freeny[-2])
ggvar_predict(x, NULL, 1:10, 0:-10, args_facet = list(scales = "free_y"))

}
