% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_predict.R
\name{ggvar_predict}
\alias{ggvar_predict}
\title{Plot VAR predictions}
\usage{
ggvar_predict(
  x,
  series = NULL,
  index_ahead,
  index_behind = NULL,
  ci = 0.95,
  args_aes = list(),
  args_line = list(),
  args_ribbon = list(fill = NA, linetype = 2, color = "blue"),
  args_labs = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{A "varest" object to get predictions from.}

\item{series}{A character vector with series (variables) to  consider. Defaults to all (\code{NULL}).}

\item{index_ahead}{A vector of labels to the x-axis, normally dates. Applied
to the predicted portion of the graph. Its length will define the
prediction horizon.}

\item{index_behind}{A vector of labels to the x-axis, normally dates. Applied
to the original portion of the graph. Its length will define the 'past'
horizon. Leave as \code{NULL} to only plot predicted values.}

\item{ci}{The confidence level for the confidence interval. Set to \code{FALSE} to omit. Used in \link[stats:predict]{predict}.}

\item{args_aes}{Defines aesthetics to differentiate the data. A named list of aesthetics names (\code{*}) -- arguments passed to \verb{ggplot2::scale_*_manual}. See more in the 'Customization' section.}

\item{args_line, args_ribbon}{Additional arguments passed to \link[ggplot2:geom_path]{geom_line} and \link[ggplot2:geom_ribbon]{geom_ribbon} (respectively). See more in the 'Customization' section.}

\item{args_labs}{Additional arguments passed to \link[ggplot2:labs]{labs}. If an empty list, will be changed to default values.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Gets predictions via \link[stats:predict]{stats::predict} and plots them. Can also plot the known
values of the past.
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

The \code{args_aes} is a list with '* = arguments to \verb{scale_*_manual} \
elements, where '*' represents the name of an aesthetic to apply to the \
data. View \code{vignette('ggplot2-specs', 'ggplot2')} to see the available \
aesthetics.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{get_gg_info}.

See \code{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function {texts_vec}. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to \link[vars:predict]{vars::predict.varest}.
}
}
}
\examples{
x <- vars::VAR(freeny[-2])
ggvar_predict(x, NULL, 1:10, 0:-10, args_facet = list(scales = "free_y"))

}
\seealso{
Other historic values plots: 
\code{\link{ggvar_fit}()},
\code{\link{ggvar_history}()}
}
\concept{historic values plots}
