% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_stability.R
\name{ggvar_stability}
\alias{ggvar_stability}
\title{Plot VAR structural stability}
\usage{
ggvar_stability(
  x,
  series = NULL,
  ci = 0.95,
  args_line = list(),
  args_hline = list(linetype = 2, color = "blue"),
  args_labs = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars:stability]{stability}, or,
directly, a "varstabil" object.}

\item{series}{A character vector with series (variables) to  consider. Defaults to all (\code{NULL}).}

\item{ci}{The confidence level for the confidence interval. Set to \code{FALSE} to omit. Used in \link[strucchange:boundary]{boundary}.}

\item{args_line, args_hline}{Additional arguments passed to \link[ggplot2:geom_path]{geom_line} and \link[ggplot2:geom_abline]{geom_hline} (respectively). See more in the 'Customization' section.}

\item{args_labs}{Additional arguments passed to \link[ggplot2:labs]{labs}. If an empty list, will be changed to default values.}

\item{args_facet}{Additional arguments passed to the faceting engine used.}

\item{...}{Arguments passed to methods, see the 'Methods' section.}
}
\value{
A \link[ggplot2:ggplot]{ggplot}.
}
\description{
Plots the result of a \link[vars:stability]{stability} call. Confidence intervals
are calculated using \link[strucchange:boundary]{boundary}.
}
\details{
\subsection{Customization}{

The graph can be customized both with the 'static' arguments passed to each layer -- using the \verb{args_*} arguments --, and, if applicable, the 'dynamic' aesthetics -- using the \code{args_aes} argument.

After built, the result can be further customized as any ggplot, adding or overwriting layers with the \link[ggplot2:gg-add]{ggplot's +}. It is useful to understand the data and the mappings coded by the package, using the function \link{get_gg_info}.

See \code{vignette('customizing-graphs')} for more details.
}

\subsection{Methods}{

The data from \code{x} is extracted with the generic function mtsdesc:::stability_setup. Each class conditions an external function to pass the \code{...} arguments to. Below there is a list with all the currently implemented classes:
\itemize{
\item Class \code{'varest'}: passed to \link[vars:stability]{vars::stability}.
\item Class \code{'varstabil'}: passed to nothing.
}
}
}
\examples{
ggvar_stability(vars::VAR(freeny[-2]))

}
\seealso{
Other model diagnostics plots: 
\code{\link{ggvar_acf}()},
\code{\link{ggvar_dispersion}()},
\code{\link{ggvar_distribution}()},
\code{\link{ggvar_history}()},
\code{\link{ggvar_select}()}
}
\concept{model diagnostics plots}
