% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varr-package.R
\docType{package}
\name{varr-package}
\alias{varr}
\alias{varr-package}
\title{varr: Utility Functions for VAR Models}
\description{
Utility functions for VAR models. Currently, ggplots for objects like like IRF's, FEVD's, predict graphs, and residual analysis. Soon to contain functions for common tests and statistics, support for other VAR-like objects, and unified builders for vector regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/ricardo-semiao/varr}
  \item \url{https://ricardo-semiao.github.io/varr/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/varr/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
